绑定了部分控件与新建了动作函数

+++ headers.h
@@ -19,7 +19,7 @@
 struct motorDetails
 {
     QString motorname;
-    double motorID;
+    int motorID;
 
     double powerControl;
